third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
(function() {
    var t = async_test('Unregistering with no existing registration');
    navigator.serviceWorker.unregister('/nothing-here/*').then(
        t.step_func(function() {
            assert_unreached('unregistering with no existing registration ' +
                             'should not succeed.');
        }),
        t.step_func(function(reason) {
            // FIXME: Make this check for NotFoundError exclusively when Chrome
            // starts generating that error
            assert_in_array(reason.name, ['UnknownError', 'NotFoundError'],
                            'unregistering with no existing registration ' +
                            'should fail with a NotFoundError');
            t.done();
        }));
}());

(function() {
    var scope = '/register-then-unregister/*';

    var t = async_test('Register then unregister');
    navigator.serviceWorker.register(
        'resources/empty-worker.js',
        { scope: scope }
    ).then(
        function(worker) {
            return navigator.serviceWorker.unregister(scope);
        },
        t.step_func(function(reason) {
            assert_unreached('registration should succeed.');
        })
    ).then(
        function() {
            t.done();
        },
        t.step_func(function(reason) {
            assert_unreached('unregistration should succeed.');
        })
    );
}());
</script>
